--- src/entities/attendance-badge/@x/badge-mock-data.ts
@@ -1,1 +1,0 @@
-export { deleteBadgeDetailOptionInMockData } from '../api/mocks/delete-badge'
