--- src/entities/attendance-badge/@x/revalidate-tag.ts
@@ -1,1 +1,0 @@
-export { ATTENDANCE_BADGE_TAG } from '../api/endpoint'
