+++ src/entities/attendance-option/@x/attendance-badge.ts
@@ -1,1 +1,2 @@
 export { type AttendanceDetailOption, AttendanceDetailOptionSchema } from '../model'
+export { type AttendanceOptionKey, AttendanceOptionKeySchema } from '../model'
