--- src/entities/attendance-option/@x/option-key.ts
@@ -1,1 +1,0 @@
-export { type AttendanceOptionKey, AttendanceOptionKeySchema } from '../model'
